limited POST,PUT,DELETE on menu

+++ src/main/java/com/propsy/backend/web/rest/MenuResource.java
@@ -1,8 +1,12 @@
 package com.propsy.backend.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
+import com.propsy.backend.domain.Authority;
 import com.propsy.backend.domain.Menu;
+import com.propsy.backend.domain.Restaurant;
+import com.propsy.backend.domain.User;
 import com.propsy.backend.repository.MenuRepository;
+import com.propsy.backend.service.UserService;
 import com.propsy.backend.web.rest.errors.BadRequestAlertException;
 import com.propsy.backend.web.rest.util.HeaderUtil;
 import io.github.jhipster.web.util.ResponseUtil;
@@ -17,6 +21,7 @@ import java.net.URISyntaxException;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 /**
  * REST controller for managing Menu.
@@ -31,8 +36,11 @@ public class MenuResource {
 
     private final MenuRepository menuRepository;
 
-    public MenuResource(MenuRepository menuRepository) {
+    private final UserService userService;
+
+    public MenuResource(MenuRepository menuRepository, UserService userService) {
         this.menuRepository = menuRepository;
+        this.userService = userService;
     }
 
     /**
@@ -46,13 +54,38 @@ public class MenuResource {
     @Timed
     public ResponseEntity<Menu> createMenu(@Valid @RequestBody Menu menu) throws URISyntaxException {
         log.debug("REST request to save Menu : {}", menu);
-        if (menu.getId() != null) {
+
+        Long menuId = menu.getId();
+        if (menuId != null) {
             throw new BadRequestAlertException("A new menu cannot already have an ID", ENTITY_NAME, "idexists");
         }
-        Menu result = menuRepository.save(menu);
-        return ResponseEntity.created(new URI("/api/menus/" + result.getId()))
-            .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
-            .body(result);
+
+        final Optional<User> isUser = userService.getUserWithAuthorities();
+
+        if(!isUser.isPresent()) {
+            log.error("User is not logged in");
+            return ResponseEntity.badRequest().build();
+        }
+
+        final User user = isUser.get();
+        String userRole = getRole(user.getAuthorities());
+
+        Restaurant restaurant = menu.getRestaurant();
+        Long menuOwnerId = null;
+
+        if(restaurant != null)
+            menuOwnerId = restaurant.getWorker().getId();
+
+        if(userRole.equals("ROLE_ADMIN") || menuOwnerId == null || menuOwnerId.equals(user.getId())) {
+            Menu result = menuRepository.save(menu);
+            return ResponseEntity.created(new URI("/api/menus/" + result.getId()))
+                .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
+                .body(result);
+        }
+
+        throw new BadRequestAlertException("Cannot create menu for a restaurant owned by a different user",
+            ENTITY_NAME,
+            "forbidden");
     }
 
     /**
@@ -68,13 +101,46 @@ public class MenuResource {
     @Timed
     public ResponseEntity<Menu> updateMenu(@Valid @RequestBody Menu menu) throws URISyntaxException {
         log.debug("REST request to update Menu : {}", menu);
-        if (menu.getId() == null) {
+
+        Long menuId = menu.getId();
+        if (menuId == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
-        Menu result = menuRepository.save(menu);
-        return ResponseEntity.ok()
-            .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, menu.getId().toString()))
-            .body(result);
+
+        final Optional<User> isUser = userService.getUserWithAuthorities();
+
+        if(!isUser.isPresent()) {
+            log.error("User is not logged in");
+            return ResponseEntity.badRequest().build();
+        }
+
+        final User user = isUser.get();
+        String userRole = getRole(user.getAuthorities());
+
+        final Optional<Menu> isMenu = menuRepository.findById(menuId);
+        if(!isMenu.isPresent()) {
+            log.error("Menu doesn't exist");
+            return ResponseEntity.badRequest().build();
+        }
+
+        final Menu dbMenu = isMenu.get();
+
+        Restaurant restaurant = dbMenu.getRestaurant();
+        Long menuOwnerId = null;
+
+        if(restaurant != null)
+            menuOwnerId = restaurant.getWorker().getId();
+
+        if(userRole.equals("ROLE_ADMIN") || menuOwnerId == null || menuOwnerId.equals(user.getId())) {
+            Menu result = menuRepository.save(menu);
+            return ResponseEntity.ok()
+                .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, menu.getId().toString()))
+                .body(result);
+        }
+
+        throw new BadRequestAlertException("Cannot update menu from a restaurant owned by a different user",
+            ENTITY_NAME,
+            "forbidden");
     }
 
     /**
@@ -115,7 +181,52 @@ public class MenuResource {
     public ResponseEntity<Void> deleteMenu(@PathVariable Long id) {
         log.debug("REST request to delete Menu : {}", id);
 
-        menuRepository.deleteById(id);
-        return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        final Optional<User> isUser = userService.getUserWithAuthorities();
+
+        if(!isUser.isPresent()) {
+            log.error("User is not logged in");
+            return ResponseEntity.badRequest().build();
+        }
+
+        final User user = isUser.get();
+        String userRole = getRole(user.getAuthorities());
+
+        final Optional<Menu> isMenu = menuRepository.findById(id);
+        if(!isMenu.isPresent()) {
+            log.error("Menu doesnt exist");
+            return ResponseEntity.badRequest().build();
+        }
+
+        final Menu menu = isMenu.get();
+
+        Restaurant restaurant = menu.getRestaurant();
+        Long menuOwnerId = null;
+
+        if(restaurant != null)
+            menuOwnerId = restaurant.getWorker().getId();
+
+        if(userRole.equals("ROLE_ADMIN") || menuOwnerId == null || menuOwnerId.equals(user.getId())) {
+            menuRepository.deleteById(id);
+            return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        }
+
+        throw new BadRequestAlertException("Cannot delete menu from a restaurant owned by a different user",
+            ENTITY_NAME,
+            "forbidden");
+    }
+
+    private String getRole(Set<Authority> authorities) {
+        String userRole = "";
+        for(Authority authority : authorities) {
+            String role = authority.getName();
+            if(role.equals("ROLE_ADMIN"))
+                userRole = role;
+            else if((role.equals("ROLE_CHEF") || role.equals("ROLE_MANAGER")) && !userRole.equals("ROLE_ADMIN"))
+                userRole = role;
+            else if(role.equals("ROLE_USER") && userRole.equals(""))
+                userRole = role;
+        }
+
+        return userRole;
     }
 }
